// tests/stores/convex/local/fixture/_generated/server.d.ts
type Env = {
  readonly CONVEX_CLOUD_URL: string;
  readonly CONVEX_SITE_URL: string;
  readonly CONNECTIONS_TEST_KEY: string | undefined;
  readonly SALESFORCE_TEST_BINDING: string | undefined;
  readonly SALESFORCE_TEST_CLIENT_ID: string | undefined;
  readonly SALESFORCE_TEST_CLIENT_SECRET: string | undefined;
  readonly SALESFORCE_TEST_CONNECTION_ID: string | undefined;
  readonly SALESFORCE_TEST_LOGIN_URL: string | undefined;
};
// ...
export declare const env: Env;